+++ extensions/api/data-management/api-configuration/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/configuration/DataManagementApiConfigurationExtension.java
@@ -17,7 +17,6 @@ package org.eclipse.dataspaceconnector.api.datamanagement.configuration;
 import org.eclipse.dataspaceconnector.api.auth.AllPassAuthenticationService;
 import org.eclipse.dataspaceconnector.api.auth.AuthenticationRequestFilter;
 import org.eclipse.dataspaceconnector.api.auth.AuthenticationService;
-import org.eclipse.dataspaceconnector.api.exception.mappers.EdcApiExceptionMapper;
 import org.eclipse.dataspaceconnector.spi.WebService;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
@@ -71,6 +70,5 @@ public class DataManagementApiConfigurationExtension implements ServiceExtension
         context.registerService(DataManagementApiConfiguration.class, new DataManagementApiConfiguration(contextAlias));
         var srv = ofNullable(service).orElse(new AllPassAuthenticationService());
         webService.registerResource(contextAlias, new AuthenticationRequestFilter(srv));
-        webService.registerResource(contextAlias, new EdcApiExceptionMapper());
     }
 }