+++ extensions/http/jersey/src/main/java/org/eclipse/dataspaceconnector/extension/jersey/JerseyConfiguration.java
@@ -18,15 +18,11 @@ import org.eclipse.dataspaceconnector.spi.EdcSetting;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 
 /**
- * Configuration class to enable or disable CORS and set various settings:
- * <ul>
- *     <li>@code edc.web.rest.cors.enabled} whether the {@link CorsFilter} is registered or not, defaults to {@code Boolean.FALSE.toString()}. So CORS is disabled by default.</li>
- *     <li>@code edc.web.rest.cors.origins} all allowed origins, defaults to {@code "*"}</li>
- *     <li>@code edc.web.rest.cors.headers} all allowed headers, defaults to {@code "origin, content-type, accept, authorization"}</li>
- *     <li>@code edc.web.rest.cors.methods} all allowed methods, defaults to {@code "GET, POST, DELETE, PUT, OPTIONS"}</li>
- * </ul>
+ * Jersey extension configuration class
  */
-public class CorsFilterConfiguration {
+public class JerseyConfiguration {
+    @EdcSetting
+    public static final String ERROR_RESPONSE_VERBOSE_SETTING = "edc.web.rest.error.response.verbose";
     @EdcSetting
     public static final String CORS_CONFIG_ORIGINS_SETTING = "edc.web.rest.cors.origins";
     @EdcSetting
@@ -39,26 +35,28 @@ public class CorsFilterConfiguration {
     private String allowedHeaders;
     private String allowedMethods;
     private boolean corsEnabled;
+    private boolean errorResponseVerbose;
 
-    private CorsFilterConfiguration() {
+    private JerseyConfiguration() {
     }
 
-    public static CorsFilterConfiguration from(ServiceExtensionContext context) {
+    public static JerseyConfiguration from(ServiceExtensionContext context) {
         var origins = context.getSetting(CORS_CONFIG_ORIGINS_SETTING, "*");
         var headers = context.getSetting(CORS_CONFIG_HEADERS_SETTING, "origin, content-type, accept, authorization");
         var allowedMethods = context.getSetting(CORS_CONFIG_METHODS_SETTING, "GET, POST, DELETE, PUT, OPTIONS");
-        var enabled = context.getSetting(CORS_CONFIG_ENABLED_SETTING, Boolean.FALSE.toString());
-        var config = new CorsFilterConfiguration();
+        var enabled = context.getSetting(CORS_CONFIG_ENABLED_SETTING, false);
+        var config = new JerseyConfiguration();
         config.allowedHeaders = headers;
         config.allowedOrigins = origins;
         config.allowedMethods = allowedMethods;
-        config.corsEnabled = Boolean.parseBoolean(enabled);
+        config.corsEnabled = enabled;
+        config.errorResponseVerbose = context.getSetting(ERROR_RESPONSE_VERBOSE_SETTING, false);
 
         return config;
     }
 
-    public static CorsFilterConfiguration none() {
-        return new CorsFilterConfiguration();
+    public static JerseyConfiguration none() {
+        return new JerseyConfiguration();
     }
 
     public String getAllowedOrigins() {
@@ -73,8 +71,11 @@ public class CorsFilterConfiguration {
         return allowedMethods;
     }
 
-
     public boolean isCorsEnabled() {
         return corsEnabled;
+    }
+
+    public boolean isErrorResponseVerbose() {
+        return errorResponseVerbose;
     }
 }