+++ extensions/http/jersey/src/main/java/org/eclipse/dataspaceconnector/extension/jersey/JerseyExtension.java
@@ -16,16 +16,17 @@ package org.eclipse.dataspaceconnector.extension.jersey;
 
 import org.eclipse.dataspaceconnector.extension.jetty.JettyService;
 import org.eclipse.dataspaceconnector.spi.WebService;
+import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
-import org.eclipse.dataspaceconnector.spi.system.Requires;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 
-@Provides({WebService.class})
-@Requires({JettyService.class})
+@Provides(WebService.class)
 public class JerseyExtension implements ServiceExtension {
     private JerseyRestService jerseyRestService;
+
+    @Inject
+    private JettyService jettyService;
 
     @Override
     public String name() {
@@ -35,13 +36,11 @@ public class JerseyExtension implements ServiceExtension {
     @Override
     public void initialize(ServiceExtensionContext context) {
         var monitor = context.getMonitor();
-        TypeManager typeManager = context.getTypeManager();
+        var typeManager = context.getTypeManager();
 
-        var jettyService = context.getService(JettyService.class);
+        var configuration = JerseyConfiguration.from(context);
 
-        var corsConfiguration = CorsFilterConfiguration.from(context);
-
-        jerseyRestService = new JerseyRestService(jettyService, typeManager, corsConfiguration, monitor);
+        jerseyRestService = new JerseyRestService(jettyService, typeManager, configuration, monitor);
 
         context.registerService(WebService.class, jerseyRestService);
     }