+++ extensions/http/jersey/src/test/java/org/eclipse/dataspaceconnector/extension/jersey/JerseyConfigurationTest.java
@@ -12,14 +12,14 @@
  *
  */
 
-package jetty;
+package org.eclipse.dataspaceconnector.extension.jersey;
 
-import org.eclipse.dataspaceconnector.extension.jersey.CorsFilterConfiguration;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.atLeastOnce;
@@ -27,16 +27,16 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class CorsFilterConfigurationTest {
+class JerseyConfigurationTest {
 
     @Test
     void ensureCorrectDefaults() {
-        ServiceExtensionContext ctx = mock(ServiceExtensionContext.class);
+        var ctx = mock(ServiceExtensionContext.class);
         var defaultValueCapture = ArgumentCaptor.forClass(String.class);
         //always return the default value
         when(ctx.getSetting(anyString(), defaultValueCapture.capture())).thenAnswer(i -> defaultValueCapture.getValue());
 
-        var config = CorsFilterConfiguration.from(ctx);
+        var config = JerseyConfiguration.from(ctx);
 
         assertThat(config.getAllowedMethods()).isEqualTo("GET, POST, DELETE, PUT, OPTIONS");
         assertThat(config.getAllowedHeaders()).isEqualTo("origin, content-type, accept, authorization");
@@ -47,13 +47,13 @@ class CorsFilterConfigurationTest {
 
     @Test
     void ensureCorrectSettings() {
-        ServiceExtensionContext ctx = mock(ServiceExtensionContext.class);
-        when(ctx.getSetting(eq(CorsFilterConfiguration.CORS_CONFIG_ENABLED_SETTING), anyString())).thenReturn("true");
-        when(ctx.getSetting(eq(CorsFilterConfiguration.CORS_CONFIG_HEADERS_SETTING), anyString())).thenReturn("origin, authorization");
-        when(ctx.getSetting(eq(CorsFilterConfiguration.CORS_CONFIG_ORIGINS_SETTING), anyString())).thenReturn("localhost");
-        when(ctx.getSetting(eq(CorsFilterConfiguration.CORS_CONFIG_METHODS_SETTING), anyString())).thenReturn("GET, POST");
+        var ctx = mock(ServiceExtensionContext.class);
+        when(ctx.getSetting(eq(JerseyConfiguration.CORS_CONFIG_ENABLED_SETTING), anyBoolean())).thenReturn(true);
+        when(ctx.getSetting(eq(JerseyConfiguration.CORS_CONFIG_HEADERS_SETTING), anyString())).thenReturn("origin, authorization");
+        when(ctx.getSetting(eq(JerseyConfiguration.CORS_CONFIG_ORIGINS_SETTING), anyString())).thenReturn("localhost");
+        when(ctx.getSetting(eq(JerseyConfiguration.CORS_CONFIG_METHODS_SETTING), anyString())).thenReturn("GET, POST");
 
-        var config = CorsFilterConfiguration.from(ctx);
+        var config = JerseyConfiguration.from(ctx);
 
         assertThat(config.getAllowedMethods()).isEqualTo("GET, POST");
         assertThat(config.getAllowedHeaders()).isEqualTo("origin, authorization");